pyproject.toml
[project]
name = "nhs-survey-dash"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "openpyxl>=3.1.5",
    "openai>=2.0.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.1.0",
]

[project.scripts]
staff-survey = "analysis.cli:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["analysis"]
